+++ Source/Mosa.Workspace.UnitTest.Debug/Mosa.Workspace.UnitTest.Debug.csproj
@@ -45,10 +45,9 @@
     <Reference Include="System" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CreateParameters.cs" />
     <Compile Include="UnitTestHelper.cs" />
     <Compile Include="UnitTest.cs" />
-    <Compile Include="UnitTests.cs" />
+    <Compile Include="UnitTestSystem.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
