+++ Source/Mosa.Workspace.UnitTest.Debug/UnitTest.cs
@@ -5,14 +5,14 @@ using System.Reflection;
 
 namespace Mosa.Workspace.UnitTest.Debug
 {
-	internal static partial class UnitTests
+	internal static partial class UnitTestSystem
 	{
 		public class UnitTest
 		{
 			public string FullMethodName { get; set; }
 			public MethodInfo Method { get; set; }
 			public MosaUnitTestAttribute UnitTestAttribute { get; set; }
-			public object[] ParameterValues { get; set; }
+			public object[] Values { get; set; }
 			public object Expected { get; set; }
 			public object Result { get; set; }
 			public bool Skipped { get; set; }
