+++ Source/Mosa.Workspace.UnitTest.Debug/UnitTestHelper.cs
@@ -6,25 +6,13 @@ namespace Mosa.Workspace.UnitTest.Debug
 {
 	public static class UnitTestHelper
 	{
-		public static object[] GetParams(this MosaUnitTestAttribute unitTest)
-		{
-			object[] values = null;
-
-			switch (unitTest.ParamCount)
-			{
-				case 1: values = new object[] { unitTest.Param1 }; break;
-				case 2: values = new object[] { unitTest.Param1, unitTest.Param2 }; break;
-				case 3: values = new object[] { unitTest.Param1, unitTest.Param2, unitTest.Param3 }; break;
-				case 4: values = new object[] { unitTest.Param1, unitTest.Param2, unitTest.Param3, unitTest.Param4 }; break;
-				case 5: values = new object[] { unitTest.Param1, unitTest.Param2, unitTest.Param3, unitTest.Param4, unitTest.Param5 }; break;
-			}
-
-			return values;
-		}
-
 		public static string ToFormattedString(this object o)
 		{
-			if (o is Byte | o is UInt16 | o is UInt32 | o is UInt64)
+			if (o is null)
+			{
+				return "NULL";
+			}
+			else if (o is Byte | o is UInt16 | o is UInt32 | o is UInt64)
 			{
 				return String.Format("0x{0:X}", o);
 			}
@@ -44,7 +32,7 @@ namespace Mosa.Workspace.UnitTest.Debug
 			return o.ToString();
 		}
 
-		public static string ToFormmattedString(this object[] objects)
+		public static string ToFormattedString(this object[] objects)
 		{
 			var sb = new StringBuilder();
